Permission for managing triggers (#842) (#922)

+++ src/trigger/trigger.controller.ts
@@ -3,7 +3,7 @@ import { ApiSecurity } from '@nestjs/swagger';
 import { PolyAuthGuard } from 'auth/poly-auth-guard.service';
 import { AuthData, AuthRequest } from 'common/types';
 import { TriggerService } from 'trigger/trigger.service';
-import { CreateTriggerDto, Role, TriggerDestination, TriggerResponseDto, TriggerSource } from '@poly/model';
+import { CreateTriggerDto, Permission, TriggerDestination, TriggerDto, TriggerResponseDto, TriggerSource, Visibility } from '@poly/model';
 import { WebhookService } from 'webhook/webhook.service';
 import { FunctionService } from 'function/function.service';
 import { AuthService } from 'auth/auth.service';
@@ -21,32 +21,43 @@ export class TriggerController {
   ) {
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Get()
   async getTriggers(@Req() req: AuthRequest) {
+    await this.authService.checkPermissions(req.user, Permission.ManageTriggers);
+
     return await this.triggerService.getTriggers(req.user.environment.id);
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Post()
   async createTrigger(@Req() req: AuthRequest, @Body() data: CreateTriggerDto) {
     const { name = null, source, destination } = data;
+
+    await this.authService.checkPermissions(req.user, Permission.ManageTriggers);
     await this.checkTriggerSource(req, source);
     await this.checkTriggerDestination(req, destination);
 
     return await this.triggerService.createTrigger(req.user.environment.id, name, source, destination);
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Get(':id')
   async getTrigger(@Req() req: AuthRequest, @Param('id') id: string) {
-    return await this.findTrigger(req.user, id);
+    const trigger = await this.findTrigger(req.user, id);
+
+    await this.checkTriggerAccess(req.user, trigger);
+
+    return trigger;
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Delete(':id')
   async deleteTrigger(@Req() req: AuthRequest, @Param('id') id: string) {
     const trigger = await this.findTrigger(req.user, id);
+
+    await this.checkTriggerAccess(req.user, trigger);
+
     return await this.triggerService.deleteTrigger(req.user.environment.id, trigger);
   }
 
@@ -92,5 +103,12 @@ export class TriggerController {
     if (!serverFunction || !await this.authService.hasEnvironmentEntityAccess(serverFunction, req.user, true)) {
       throw new NotFoundException(`Server function ${functionId} not found`);
     }
+  }
+
+  private async checkTriggerAccess(authData: AuthData, trigger: TriggerDto) {
+    await this.authService.checkEnvironmentEntityAccess({
+      ...trigger,
+      visibility: Visibility.Environment,
+    }, authData, false, Permission.ManageTriggers);
   }
 }
